Restructure CirclePacking as a self-contained chart
Applies the same restructuring as Treemap/Dendrogram/RadialDendrogram:
- The old plot-only CirclePacking is renamed CirclePackingPlot; the
  new CirclePacking is a self-contained chart that renders
  <RectangularChart><CirclePackingPlot/></RectangularChart>
  internally, so consumers no longer need to wrap it themselves.
- CirclePackingBase pulls its plot dimensions directly from
  chartSelectors.dimensions.plot.* via useSelector instead of
  withRectangularPlot-injected props.
- colorHierarchyNode calls now pass the theme's background, so
  deeper nodes darken (rather than brighten) against a dark theme.

Docs and stories updated to match. Verified visually via a built
Storybook: zoomed in an extra level past the story's own play()
interaction with no NaN positions and correct breadcrumb state.
