+++ packages/react/src/lib/components/Plots/CirclePacking/__image_snapshots__/circle-packing-plot-unit-tsx-circle-packing-plot-using-canvas-should-render-correctly-1-snap.png
