+++ packages/react/src/lib/components/Plots/CirclePacking/index.ts
@@ -1,1 +1,2 @@
 export * from "./CirclePacking";
+export * from "./CirclePackingPlot";
